+++ src/main/java/fll/scheduler/SchedulerUI.java
@@ -31,7 +31,6 @@ import java.io.OutputStreamWriter;
 import java.io.Reader;
 import java.io.Writer;
 import java.net.URL;
-import java.sql.SQLException;
 import java.text.ParseException;
 import java.time.LocalTime;
 import java.util.Collection;
@@ -90,13 +89,11 @@ import static org.checkerframework.checker.nullness.util.NullnessUtil.castNonNul
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import fll.Team;
 import fll.Utilities;
-import fll.db.CategoryColumnMapping;
 import fll.db.TournamentParameters;
 import fll.scheduler.SchedParams.InvalidParametersException;
 import fll.scheduler.TournamentSchedule.ColumnInformation;
 import fll.util.CellFileReader;
 import fll.util.ExcelCellReader;
-import fll.util.FLLInternalException;
 import fll.util.FLLRuntimeException;
 import fll.util.FormatterUtils;
 import fll.util.GuiExceptionHandler;
@@ -104,8 +101,6 @@ import fll.util.ProgressDialog;
 import fll.web.StoreColumnNames;
 import fll.xml.ChallengeDescription;
 import fll.xml.ChallengeParser;
-import fll.xml.ScoreCategory;
-import fll.xml.SubjectiveScoreCategory;
 import net.mtu.eggplant.util.BasicFileFilter;
 
 /**
@@ -323,7 +318,6 @@ public class SchedulerUI extends JFrame {
     scheduleMenu.add(mOpenScheduleAction);
     scheduleMenu.add(mReloadFileAction);
     scheduleMenu.add(mRunOptimizerAction);
-    scheduleMenu.add(mWriteSchedulesAction);
     scheduleMenu.add(mDisplayGeneralScheduleAction);
     scheduleMenu.add(saveScheduleAction);
     menubar.add(scheduleMenu);
@@ -331,7 +325,6 @@ public class SchedulerUI extends JFrame {
     setJMenuBar(menubar);
 
     // initial state
-    mWriteSchedulesAction.setEnabled(false);
     mDisplayGeneralScheduleAction.setEnabled(false);
     mRunOptimizerAction.setEnabled(false);
     mReloadFileAction.setEnabled(false);
@@ -781,8 +774,7 @@ public class SchedulerUI extends JFrame {
     return toolbar;
   }
 
-  @RequiresNonNull({ "mScheduleFilename", "mOpenScheduleAction", "mReloadFileAction", "mWriteSchedulesAction",
-                     "mDisplayGeneralScheduleAction" })
+  @RequiresNonNull({ "mScheduleFilename", "mOpenScheduleAction", "mReloadFileAction", "mDisplayGeneralScheduleAction" })
   private JToolBar createScheduleToolbar(@UnderInitialization(JFrame.class) SchedulerUI this) {
     final JToolBar toolbar = new JToolBar("Schedule Toolbar");
     toolbar.setFloatable(false);
@@ -791,7 +783,6 @@ public class SchedulerUI extends JFrame {
     toolbar.addSeparator();
     toolbar.add(mOpenScheduleAction);
     toolbar.add(mReloadFileAction);
-    toolbar.add(mWriteSchedulesAction);
     toolbar.add(mDisplayGeneralScheduleAction);
 
     return toolbar;
@@ -964,206 +955,6 @@ public class SchedulerUI extends JFrame {
 
       } else {
         // user canceled
-        return;
-      }
-    }
-  }
-
-  private final Action mWriteSchedulesAction = new WriteSchedulesAction();
-
-  private final class WriteSchedulesAction extends AbstractAction {
-    WriteSchedulesAction() {
-      super("Write Detailed Schedules");
-      putValue(SMALL_ICON, Utilities.getIcon("toolbarButtonGraphics/general/Export16.gif"));
-      putValue(LARGE_ICON_KEY, Utilities.getIcon("toolbarButtonGraphics/general/Export24.gif"));
-      putValue(SHORT_DESCRIPTION, "Write out the detailed schedules as a PDF");
-      // putValue(MNEMONIC_KEY, KeyEvent.VK_X);
-    }
-
-    @Override
-    public void actionPerformed(final ActionEvent ae) {
-      try {
-        final File directory = getScheduleFile().getParentFile();
-        if (null == directory) {
-          final String message = "Cannot find a parent directory for "
-              + getScheduleFile()
-              + ", skipping writing of schedules";
-          LOGGER.warn("Cannot find a parent directory for {}, skipping writing of schedules", getScheduleFile());
-          JOptionPane.showMessageDialog(SchedulerUI.this, "Schedules not written", message,
-                                        JOptionPane.WARNING_MESSAGE);
-          return;
-        }
-
-        final String baseFilename = tournamentName;
-        LOGGER.info("Writing detailed schedules to "
-            + directory.getAbsolutePath());
-
-        final SwingWorker<Boolean, Void> detailedSchedulesTask = new SwingWorker<Boolean, Void>() {
-          @Override
-          protected Boolean doInBackground() {
-            try {
-              // output the schedule as a CSV file so that table optimization information can
-              // be saved
-              final File fullSchedule = new File(directory, String.format("%s-schedule.csv", baseFilename));
-              try (OutputStream os = new FileOutputStream(fullSchedule)) {
-                getScheduleData().outputScheduleAsCSV(os);
-              }
-
-              getScheduleData().outputDetailedSchedules(directory, baseFilename);
-              return true;
-            } catch (final IOException e) {
-              final Formatter errorFormatter = new Formatter();
-              errorFormatter.format("Error writing detailed schedules: %s", e.getMessage());
-              LOGGER.error(errorFormatter, e);
-              JOptionPane.showMessageDialog(SchedulerUI.this, errorFormatter, "Error", JOptionPane.ERROR_MESSAGE);
-              return false;
-            }
-          }
-
-          @Override
-          protected void done() {
-            progressDialog.setVisible(false);
-          }
-        };
-        // make sure the task doesn't start until the window is up
-        progressDialog.addComponentListener(new ComponentAdapter() {
-          @Override
-          public void componentShown(final ComponentEvent e) {
-            progressDialog.removeComponentListener(this);
-            detailedSchedulesTask.execute();
-          }
-        });
-        progressDialog.setVisible(true);
-        if (!detailedSchedulesTask.get()) {
-          // an error occurred
-          return;
-        }
-
-        final int answer = JOptionPane.showConfirmDialog(SchedulerUI.this,
-                                                         "Would you like to write out the score sheets as well?",
-                                                         "Write Out Scoresheets?", JOptionPane.YES_NO_OPTION);
-        if (JOptionPane.YES_OPTION != answer) {
-          return;
-        }
-
-        if (null == challengeDescription) {
-          promptForChallengeDescription();
-        }
-        final ChallengeDescription description = challengeDescription;
-        if (null == description) {
-          return;
-        }
-
-        final SwingWorker<Boolean, Void> performanceSheetsTask = new SwingWorker<Boolean, Void>() {
-          @Override
-          protected Boolean doInBackground() {
-            final File scoresheetFile = new File(directory, baseFilename
-                + "-scoresheets.pdf");
-            try (FileOutputStream scoresheetFos = new FileOutputStream(scoresheetFile)) {
-
-              getScheduleData().outputPerformanceSheets(tournamentName, scoresheetFos, description);
-              return true;
-            } catch (IOException | SQLException e) {
-              final Formatter errorFormatter = new Formatter();
-              errorFormatter.format("Error writing performance sheets: %s", e.getMessage());
-              LOGGER.error(errorFormatter, e);
-              JOptionPane.showMessageDialog(SchedulerUI.this, errorFormatter, "Error", JOptionPane.ERROR_MESSAGE);
-              return false;
-            }
-          }
-
-          @Override
-          protected void done() {
-            progressDialog.setVisible(false);
-          }
-        };
-        // make sure the task doesn't start until the window is up
-        progressDialog.addComponentListener(new ComponentAdapter() {
-          @Override
-          public void componentShown(final ComponentEvent e) {
-            progressDialog.removeComponentListener(this);
-            performanceSheetsTask.execute();
-          }
-        });
-        progressDialog.setVisible(true);
-        if (!performanceSheetsTask.get()) {
-          // an error occurred
-          return;
-        }
-
-        if (null == columnInfo) {
-          throw new IllegalStateException("No column information");
-        }
-        final Collection<CategoryColumnMapping> categoryMappings = columnInfo.getSubjectiveColumnMappings();
-        final MapSubjectiveHeaders mapDialog = new MapSubjectiveHeaders(SchedulerUI.this, description);
-        mapDialog.setLocationRelativeTo(SchedulerUI.this);
-        mapDialog.setVisible(true);
-        if (mapDialog.isCanceled()) {
-          return;
-        }
-
-        final Map<ScoreCategory, Collection<String>> categoryToScheduleColumns = new HashMap<>();
-        final Map<ScoreCategory, @Nullable String> filenameSuffixes = new HashMap<>();
-        for (final SubjectiveScoreCategory scoreCategory : description.getSubjectiveCategories()) {
-          final Collection<CategoryColumnMapping> scheduleColumnMappings = categoryMappings.stream()
-                                                                                           .filter(mapping -> mapping.getCategoryName()
-                                                                                                                     .equals(scoreCategory.getName()))
-                                                                                           .toList();
-          if (scheduleColumnMappings.isEmpty()) {
-            throw new FLLInternalException("Did not find a schedule column for "
-                + scoreCategory.getTitle());
-          }
-          final Collection<String> scheduleColumns = scheduleColumnMappings.stream()
-                                                                           .map(CategoryColumnMapping::getScheduleColumn)
-                                                                           .toList();
-          categoryToScheduleColumns.put(scoreCategory, scheduleColumns);
-          filenameSuffixes.put(scoreCategory, mapDialog.getFilenameSuffixForCategory(scoreCategory));
-        }
-
-        final SwingWorker<Boolean, Void> subjectiveSheetsTask = new SwingWorker<Boolean, Void>() {
-          @Override
-          protected Boolean doInBackground() {
-            try {
-              getScheduleData().outputSubjectiveSheets(tournamentName, directory.getAbsolutePath(), baseFilename,
-                                                       description, categoryToScheduleColumns, filenameSuffixes);
-              return true;
-            } catch (final IOException e) {
-              final Formatter errorFormatter = new Formatter();
-              errorFormatter.format("Error writing subjective sheets: %s", e.getMessage());
-              LOGGER.error(errorFormatter, e);
-              JOptionPane.showMessageDialog(SchedulerUI.this, errorFormatter, "Error", JOptionPane.ERROR_MESSAGE);
-              return false;
-            }
-          }
-
-          @Override
-          protected void done() {
-            progressDialog.setVisible(false);
-          }
-        };
-        // make sure the task doesn't start until the window is up
-        progressDialog.addComponentListener(new ComponentAdapter() {
-          @Override
-          public void componentShown(final ComponentEvent e) {
-            progressDialog.removeComponentListener(this);
-            subjectiveSheetsTask.execute();
-          }
-        });
-        progressDialog.setVisible(true);
-        if (!subjectiveSheetsTask.get()) {
-          // an error occurred
-          return;
-        }
-
-        JOptionPane.showMessageDialog(SchedulerUI.this, "See '"
-            + directory.getAbsolutePath()
-            + "' for all outputs", "Information", JOptionPane.INFORMATION_MESSAGE);
-
-      } catch (InterruptedException | ExecutionException e) {
-        final Formatter errorFormatter = new Formatter();
-        errorFormatter.format("Unexpected error generating output: %s", e.getMessage());
-        LOGGER.error(errorFormatter, e);
-        JOptionPane.showMessageDialog(SchedulerUI.this, errorFormatter, "Error", JOptionPane.ERROR_MESSAGE);
         return;
       }
     }
@@ -1368,7 +1159,6 @@ public class SchedulerUI extends JFrame {
           + ":"
           + mScheduleSheetName);
 
-      mWriteSchedulesAction.setEnabled(true);
       mDisplayGeneralScheduleAction.setEnabled(true);
       mRunOptimizerAction.setEnabled(true);
       mReloadFileAction.setEnabled(true);