+++ src/main/java/fll/web/report/PitSigns.java
@@ -120,7 +120,8 @@ public class PitSigns extends BaseFLLServlet {
       // then render all pit signs
       final @Nullable String teamNumberStr = request.getParameter("team_number");
       if (null == teamNumberStr) {
-        response.setHeader("Content-Disposition", "filename=pit_signs.pdf");
+        response.setHeader("Content-Disposition",
+                           String.format("attachment; filename=\"%s_pit_signs.pdf\"", tournament.getName()));
 
         for (final TournamentTeam team : Queries.getTournamentTeams(connection, tournament.getTournamentID())
                                                 .values()) {
@@ -131,7 +132,8 @@ public class PitSigns extends BaseFLLServlet {
         }
       } else {
         final int teamNumber = Integer.parseInt(teamNumberStr);
-        response.setHeader("Content-Disposition", String.format("filename=pit_sign_%d.pdf", teamNumber));
+        response.setHeader("Content-Disposition", String.format("attachment; filename=\"%s_pit_sign_%d.pdf\"",
+                                                                tournament.getName(), teamNumber));
 
         final TournamentTeam team = TournamentTeam.getTournamentTeamFromDatabase(connection, tournament, teamNumber);
 
@@ -203,7 +205,7 @@ public class PitSigns extends BaseFLLServlet {
     final Element judgingGroupBlock = FOPUtils.createXslFoElement(document, FOPUtils.BLOCK_TAG);
     page.appendChild(judgingGroupBlock);
     judgingGroupBlock.appendChild(document.createTextNode(String.format("Judging Group: %s", team.getJudgingGroup())));
-    judgingGroupBlock.setAttribute("font-size",  "22pt");
+    judgingGroupBlock.setAttribute("font-size", "22pt");
     final @Nullable String wave = team.getWave();
     if (null != wave) {
       final Element waveText = FOPUtils.createXslFoElement(document, FOPUtils.INLINE_TAG);